Version 2.0
Uploading Version 2.0 This has been created using PHP. It converts passwords to a MD5 hash, then truncates it and appends a prefix. After which it is then converted to bcrypt. This make it very difficult to crack this hash type.

This includes 2 functions, one for generating the hashes, and a second for verifying that 2 hashes match. This is required for bcrypt as the same input can generate different hashes.